% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_gel_wells.R
\name{assign_gel_wells}
\alias{assign_gel_wells}
\title{assign_gel_wells}
\usage{
assign_gel_wells(filepath)
}
\arguments{
\item{filepath}{= filepath to excel file containing data exported from GelAnalyzer Software.
The file must contain a column labeled "Gel_Section" that specifies which section of the gel the data refers to.}
}
\value{
= tibble containing data with Well ID.
}
\description{
This function is used to format data exported from the band detector software Gel Analyzer (http://www.gelanalyzer.com/) when up to 2x 96 well plates are run on a gel.
Lanes 1 and 50 of each section should always contain the molecular weight marker.
The lanes of the 96 well gel are always filled in a certain pattern that allows us to determine the Well ID corresponding to a given lane.
}
\examples{


d1 = assign_gel_wells("/filepath_of_data_exported_from_Gel_Analyzer.xlsx")


}
