% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtwarper.R
\name{dtwarper}
\alias{dtwarper}
\title{dtwarper: Takes a data frame containing time series data and applies dynamic time warping.}
\usage{
dtwarper(data, scale = TRUE)
}
\value{
a distance matrix.
}
\description{
This function generates a distance matrix using dynamic time warping as implemented in the TSclust package.
}
\examples{
# Data frame called data, where each column contains time series data for a different feature ID
dist = dtwarper(data, scale = TRUE)

}
