% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_R2.R
\name{facet_R2}
\alias{facet_R2}
\title{facet_R2}
\usage{
facet_R2(df, y, x, by)
}
\arguments{
\item{df}{data frame}

\item{y}{y axis name}

\item{x}{x axis name}

\item{by}{name of the factor used to split the data into facets.}
}
\value{
a data frame containing the R^2 values for the linear model from each facet.
}
\description{
This function is used to calculate the R^2 values for different facets.
This makes it easy to add R^2 values to different facets of a ggplot using the geom_text function.
}
\examples{

data(mtcars)

R2 = facet_R2(mtcars,y = "mpg",x = "hp",by = "gear")

p1 = ggplot(mtcars,aes(mpg,hp))+
geom_point()+
geom_smooth(method = "lm")+
facet_grid(~gear)+
geom_text(data = R2,aes(x = 30, y = 300,label = R2))

p1

}
