% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_n_similar.R
\name{find_n_similar}
\alias{find_n_similar}
\title{find_n_similar: Find the most similar patterns to a pattern of interest.}
\usage{
find_n_similar(dist, POI_name, n = 10)
}
\value{
: A vector of IDs sorted by their similarity to the POI: pattern of interest
}
\description{
This function is designed to work in combination with dtwarper.
Dtwarper is intended to generate a distance matrix using dynamic time warping, which is then leveraged by find_n_similar() to find n IDs that produce the most similar patterns.
}
\examples{
# dist generated with dtwarper is used to find IDs producing the 25 most similar patterns to that of the "interesting pattern".
IDs = find_n_similar(dist,interesting_pattern, n = 25)
}
