% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CFU.R
\name{format_CFU}
\alias{format_CFU}
\title{format_CFU
This function is used to parse an excel template containing colony counts and metadata corresponding to CFU spots on agar plates and assign the appropriate sample.IDs.
Note that this function assumes that 2uL is plated for each spot.}
\usage{
format_CFU(Filepath, Dilution_orientation = "Row")
}
\arguments{
\item{Filepath}{= The file path to the filled out excel template containing the colony counts and metadata . This must be the specific excel template designed for use with this function}

\item{Dilution_orientation}{= Specify whether the dilution is oriented row wise or plate wise.}
}
\value{
= a data frame with the calculated CFU/mL for each Sample.ID
}
\description{
format_CFU
This function is used to parse an excel template containing colony counts and metadata corresponding to CFU spots on agar plates and assign the appropriate sample.IDs.
Note that this function assumes that 2uL is plated for each spot.
}
\examples{
df = format_CFU(Filepath = "/filepath",Dilution_orientation = "Row")

}
