% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.in.R
\name{\%nin\%}
\alias{\%nin\%}
\title{The not-in operator for R.}
\usage{
x \%nin\% y
}
\value{

}
\description{
The not-in operator for R.
}
\examples{

# Will return rows where carb isn't equal to 1 or 2
d1 = dplyr::filter(mtcars,carb \%in\% c(1,2))

}
