% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nxp_norm.R
\name{nxp_norm}
\alias{nxp_norm}
\title{nxp_norm: Generates an excel file with the info needed to normalize each well of a 96 well plate to a specified concentration at a given volume.
The resulting file is intended to serve as input to the Biomek NXP.}
\usage{
nxp_norm(
  filepath,
  dilution_factor,
  target_concentration = 5 * 10^7,
  final_volume = 200,
  output_filepath
)
}
\arguments{
\item{filepath}{= filepath to flow cytometry results. Note this needs to be a csv file with a column called Gate conatining a Gate named Bacteria.}

\item{dilution_factor}{= the dilution factor that the flow cytometry wells found at the specified filepath were diluted by.}

\item{target_concentration}{= the target concentration (cells/mL) you would like in the final solution.}

\item{final_volume}{= the final volume in uL that you would like in each well.}

\item{output_filepath}{= the filepath where you would like the resulting file to be written to.}
}
\value{

}
\description{
Notes:
1. This file format was designed to be used with an existing Biomek NXP protocol for gDNA normalization. That is why the collum headers don't indicate what is really in the well.
In the future I may go back and fix this to be not so confusing, this was done to be a quick hacky fix.
2. For some reason the biomek NXP seems to allow only whole numbers. If decimal places are found it just ignores the entire number.
}
