% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_clbP.R
\name{parse_clbP}
\alias{parse_clbP}
\title{parse_clbP}
\usage{
parse_clbP(biotek_export, Plate_Type = 96)
}
\arguments{
\item{Plate_Type}{}
}
\value{
a tidy data frame with OD600 and fluorescence from activity based probe
}
\description{
This function takes the exported .txt file from the clbP_probe protocol on a biotek plate reader and parses the data into a tidy data frame containg the OD600 and fluorescence,
}
