% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gas_data.R
\name{parse_gas_data}
\alias{parse_gas_data}
\title{parse_gas_data}
\usage{
parse_gas_data(filepath)
}
\arguments{
\item{filepath}{The file path of the exported file from the Ankom Gas System software. Should be an .xls file}
}
\value{
a tibble containing the parsed data
}
\description{
This function is intended to parse the exported data from the Ankom Gas System software into a format that is easy to work with.
}
\examples{
formatted_data = parse_gas_data("filepath_of_the_xls_file_exported_by_GPM_software")
}
