% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_leading_0.R
\name{rm_leading_0}
\alias{rm_leading_0}
\title{rm_leading_0}
\usage{
rm_leading_0(Wells)
}
\arguments{
\item{Wells:}{vector of well IDs}
}
\value{
vector of well IDs modified to remove the leading 0.
}
\description{
useful function to remove leading 0s from well IDS.
This can be usefull when trying to map data from sample.IDs with leading 0s to other data without leading 0s.
Created it as a function so I don't have to look up the regular expression every time I use it.
}
\examples{
Wells = c("A01","A02", "B04")

rm_leading_0(Wells)
##Output###
"A1","A2","B4"
}
